+++ lib/quarter.mli
@@ -4,4 +4,4 @@ val compare : t -> t -> int
 val pp : t Fmt.t
 val of_string : string -> (t option, [ `Msg of string ]) result
 val of_filename : filename:string -> t option
-val check : t option -> t option -> bool
+val check : t option -> starts:t option -> ends:t option -> bool