+++ test/unit/test_filter.ml
@@ -35,15 +35,24 @@ let id2 = "Id2"
 let id3 = "ID3"
 
 let kr1 =
-  let kind = KR.Kind.Work (KR.Work.v ~title:t1 ~id:New_KR ~quarter:None) in
+  let kind =
+    KR.Kind.Work
+      (KR.Work.v ~title:t1 ~id:New_KR ~start_quarter:None ~end_quarter:None)
+  in
   KR.v ~kind ~project:p1 ~objective:o1 ~time_entries:te1 []
 
 let kr2 =
-  let kind = KR.Kind.Work (KR.Work.v ~title:t2 ~id:(ID id2) ~quarter:None) in
+  let kind =
+    KR.Kind.Work
+      (KR.Work.v ~title:t2 ~id:(ID id2) ~start_quarter:None ~end_quarter:None)
+  in
   KR.v ~kind ~project:p2 ~objective:o2 ~time_entries:te2 []
 
 let kr3 =
-  let kind = KR.Kind.Work (KR.Work.v ~title:t3 ~id:(ID id3) ~quarter:None) in
+  let kind =
+    KR.Kind.Work
+      (KR.Work.v ~title:t3 ~id:(ID id3) ~start_quarter:None ~end_quarter:None)
+  in
   KR.v ~kind ~project:p2 ~objective:o2 ~time_entries:te3 []
 
 let report () = fst @@ Okra.Report.of_krs [ kr1; kr2; kr3 ]